Start using size_t for length of buffer more

+++ libretro/dosio.c
@@ -62,7 +62,7 @@ void* file_create(const char *filename)
 	return ret;
 }
 
-size_t file_lread(void* handle, void *data, DWORD length)
+size_t file_lread(void* handle, void *data, size_t length)
 {
 	size_t readsize;
 	if (read_file(handle, data, length, &readsize) == 0)
@@ -70,7 +70,7 @@ size_t file_lread(void* handle, void *data, DWORD length)
 	return readsize;
 }
 
-size_t file_lwrite(void *handle, void *data, DWORD length)
+size_t file_lwrite(void *handle, void *data, size_t length)
 {
 	size_t writesize;
 	if (write_file(handle, data, length, &writesize) == 0)
@@ -78,7 +78,7 @@ size_t file_lwrite(void *handle, void *data, DWORD length)
 	return writesize;
 }
 
-size_t file_read(void *handle, void *data, WORD length)
+size_t file_read(void *handle, void *data, size_t length)
 {
 	size_t readsize;
 	if (read_file(handle, data, length, &readsize) == 0)
@@ -86,7 +86,7 @@ size_t file_read(void *handle, void *data, WORD length)
 	return readsize;
 }
 
-size_t file_write(void *handle, void *data, WORD length)
+size_t file_write(void *handle, void *data, size_t length)
 {
 	size_t writesize;
 	if (write_file(handle, data, length, &writesize) == 0)
